User story 5 & 6 added
