+++ README.md
@@ -26,6 +26,53 @@ Functionalities requested:
 -  You only can peep if you are logged in.
 
 
+Setting up testing and development environments:
+------------------------------------------------
+
+In order to play around with this app you need to clone the repo shown below and execute the following commands to be complete functional.
+
+```sh
+$ git clone git@github.com:Willibaur/chitter-challenge.git
+$ bundle
+```
+
+You need to create two databases for test and development environments, I used psql where you can execute
+
+```sh
+$ psql
+$ create database chitter-test;
+$ create database chitter-development;
+```
+
+Now you can run a rakefile included in the repo for db admin functionalities
+
+```sh
+$ rake auto_migrate                  # For development db purposes
+$ rake auto_migrate RACK_ENV=test    # For testing db purposes
+```
+
+For inspecting all tests written previously you need to run
+
+```sh
+$ rspec
+```
+
+Lastly, if you want to run a development environment for which I used rackup, you only need to run
+
+```sh
+$ rackup
+```
+
+Web access application
+-----------------------
+
+[![Build Status](https://travis-ci.org/makersacademy/chitter-challenge.svg?branch=master)](https://travis-ci.org/makersacademy/chitter-challenge)
+
+This app was deployed to the web thanks to Heroku cloud service, it is called:
+
+### [Chitter](http://chitter-willibaur.herokuapp.com)
+
+
 Contributors
 ------------
