+++ app/app.rb
@@ -2,80 +2,10 @@ ENV['RACK_ENV'] ||= 'development'
 
 require 'sinatra/base'
 require 'sinatra/flash'
+require 'sinatra/partial'
 
+require_relative 'controllers/peeps'
+require_relative 'controllers/sessions'
+require_relative 'controllers/users'
 require_relative 'data_mapper_setup'
-
-class ChitterChallenge < Sinatra::Base
-
-  register Sinatra::Flash
-
-  use Rack::MethodOverride
-
-  enable :sessions
-  set :session_secret, 'super secret'
-
-  helpers do
-    def current_user
-      @current_user ||= User.get(session[:user_id])
-    end
-  end
-
-  get '/' do
-    'Welcome to Chitter'
-  end
-
-  get '/users/new' do
-    erb :'users/new'
-  end
-
-  post '/users' do
-    @user = User.new(name: params[:name],
-                     username: params[:username],
-                     email: params[:email],
-                     password: params[:password],
-                     password_confirmation: params[:password_confirmation])
-    if @user.save
-      session[:user_id] = @user.id
-      redirect '/peeps'
-    else
-      flash.now[:errors] = @user.errors.full_messages
-      erb :'users/new'
-    end
-  end
-
-  get '/sessions/new' do
-    erb :'sessions/new'
-  end
-
-  post '/sessions' do
-    user = User.authenticate(params[:email], params[:password])
-    if user
-      session[:user_id] = user.id
-      redirect '/peeps'
-    else
-      flash.now[:errors] = ['The email or password is incorrect']
-      erb :'sessions/new'
-    end
-  end
-
-  delete '/sessions' do
-    session[:user_id] = nil
-    flash.keep[:notice] = 'goodbye!'
-    redirect '/peeps'
-  end
-
-
-  get '/peeps/new' do
-    erb :'peeps/new'
-  end
-
-  get '/peeps' do
-    @peeps = Peep.all
-    erb :'peeps/index'
-  end
-
-  post '/peeps' do
-    current_user.peeps.create(message: params[:message])
-    redirect '/peeps'
-  end
-end
+require_relative 'server'
