+++ app/models/peep.rb
@@ -1,8 +1,9 @@
 class Peep
   include DataMapper::Resource
 
-  property :id,      Serial
-  property :message, String
+  property :id,         Serial
+  property :message,    String
+  property :created_at, Time
 
   belongs_to :user
 end
