+++ app/views/_flash.erb
@@ -1,0 +1,15 @@
+<div>
+  <% if flash[:notice] %>
+  <%= flash[:notice] %>
+  <% end %>
+</div>
+
+<div>
+  <% if flash[:errors] && !flash[:errors].empty? %>
+    <ul id='errors'>
+      <%= flash[:errors].each do |error| %>
+        <li> <%= error %>  </li>
+      <% end %>
+    </ul>
+  <% end %>
+</div>
